[#11] Add Enum field and improve typehint check
An enum field can accept any Enum or values from
a specific Enum class.

Fields now check the typehint used in the class is correct
for the types the field supports.
If a typehint is used with a field that does not support it
a TypeError will be raised, but because this check happens
in "__set_name__" the actual error is a RuntimeError.